fix(studylog): change column type to int from bigint

+++ backend/src/main/resources/db/migration/prod/V11__add_views_studylog_table.sql
@@ -1,1 +1,1 @@
-ALTER TABLE studylog ADD views BIGINT DEFAULT 0;
+ALTER TABLE studylog ADD views INT DEFAULT 0;
